// frontend/src/lib/api/client/types.gen.ts
export type NotificationRead = {
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at?: string | null;
    /**
     * Id
     */
    id: string;
    /**
     * User Id
     */
    user_id: string;
    /**
     * Type
     */
    type: string;
    /**
     * Title
     */
    title: string;
    /**
     * Body
     */
    body?: string | null;
    /**
     * Payload
     */
    payload?: {
        [key: string]: unknown;
    } | null;
    /**
     * Is Read
     */
    is_read: boolean;
};